app/requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
attrs
backoff
certifi
charset-normalizer
colorama
contourpy
cycler
dataclasses-json
defusedxml
fonttools
frozenlist
idna
inference
kiwisolver
marshmallow
matplotlib
multidict
mypy_extensions
numpy
opencv-python
packaging
pillow
propcache
py-cpuinfo
pyparsing
python-dateutil
python-dotenv
PyYAML
requests
scipy
six
supervision
tqdm
typing-inspect
typing_extensions
urllib3
yarl
flask
fastapi
uvicorn
python-multipart
